优化

+++ AutomaticTask/entity/config/table_task_time.go
@@ -5,6 +5,6 @@ type TableTaskTime struct {
 	Date      string `db:"date" json:"date" field:"date"`
 	StartTime string `db:"start_time" json:"startTime" field:"start_time"`
 	EndTime   string `db:"end_time" json:"endTime" field:"end_time"`
-	IsScan    int8   `db:"is_scan" json:"isScan" field:"is_scan"`
-	State     int8   `db:"state" json:"state" field:"state"`
+	IsScan    int    `db:"is_scan" json:"isScan" field:"is_scan"`
+	State     int    `db:"state" json:"state" field:"state"`
 }
